Update requirements + fix fallout from that
I haven't done this in a long time, so there are a bunch of changes.
Most notably:

Upgrading pelican means that a new way of running the dev server is
recommended. A new `Makefile` has been added, borrowed from a
`pelican-quickstart` project.

pelica-jupyter was not working with nbconvert>=6.0, so nbconvert was
downgraded